requirements.txt
alembic==1.0.0
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.4
bcrypt==3.1.4
cffi==1.11.5
click==6.7
cryptography==2.3.1
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.4.0
Flask-Cors==3.0.4
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SendGrid==0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==5.0.0
pluggy==0.8.1
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
py==1.7.0
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.6.4
pylint==2.2.2
pylint-flask==0.5
pylint-plugin-utils==0.4
pytest==4.1.1
python-dateutil==2.7.3
python-editor==1.0.3
python-http-client==3.1.0
sendgrid==5.3.0
six==1.11.0
SQLAlchemy==1.2.12
Werkzeug==0.14.1
wrapt==1.11.0

